Refactor Config
